+++ platypush/plugins/tts/google/__init__.py
@@ -33,9 +33,10 @@ class TtsGooglePlugin(TtsPlugin):
             stored, see https://cloud.google.com/text-to-speech/docs/basics.
         :param kwargs: Extra arguments to be passed to the
             :class:`platypush.plugins.tts.TtsPlugin` constructor, including
-            ``output_device``. ``output_device`` accepts a
+            ``output_device`` and ``output_volume``. ``output_device`` accepts a
             PortAudio/sounddevice device index, PortAudio/sounddevice device
             name, or PulseAudio/PipeWire sink name (requires ``pactl``).
+            ``output_volume`` is a playback volume percentage.
         """
         super().__init__(**kwargs)