+++ platypush/plugins/tts/piper/__init__.py
@@ -72,9 +72,10 @@ class TtsPiperPlugin(TtsPlugin):
             speech on some audio backends (default: 1).
         :param kwargs: Extra arguments to be passed to the
             :class:`platypush.plugins.tts.TtsPlugin` constructor, including
-            ``output_device``. ``output_device`` accepts a
+            ``output_device`` and ``output_volume``. ``output_device`` accepts a
             PortAudio/sounddevice device index, PortAudio/sounddevice device
             name, or PulseAudio/PipeWire sink name (requires ``pactl``).
+            ``output_volume`` is a playback volume percentage.
         """
         from piper import PiperVoice